MLflow/DPT/requirements.txt
chromadb==0.3.27
discord==2.3.1
openai==0.27.8
python-dotenv==1.0.0
GitPython==3.1.32
mlflow==2.4.2
